refactor(molecules): drop ShareButton/ShareModal, EventView composes ShareContent
Renames EventShare to ShareContent (folder = primary export), keeping the
ShareContent block and its private CopyField and deleting the ShareButton
trigger + private ShareModal. EventView now renders the share trigger and
wraps ShareContent in the trigger-based Modal with the "invite a friend"
header — reproducing today's dialog. RegistrationForm's thank-you screen
still reuses ShareContent. Barrel drops ShareButton/ShareButtonProps.

Part of #25
