--- src/components/molecules/EventShare/index.ts
@@ -1,2 +1,0 @@
-export { ShareButton, ShareContent } from './EventShare'
-export type { ShareButtonProps, ShareContentProps } from './EventShare'
