+++ src/components/molecules/ShareContent/index.ts
@@ -1,0 +1,2 @@
+export { ShareContent } from './ShareContent'
+export type { ShareContentProps } from './ShareContent'
